refactor: Extract SelectDatasetStep from configurator
