+++ app/configurator/components/configurator.tsx
@@ -1,40 +1,23 @@
-import { filter } from "lodash";
-import React, { useRef, useState } from "react";
+import React from "react";
 import { Box, Link, LinkProps } from "theme-ui";
-import { useDebounce } from "use-debounce";
 import { useConfiguratorState } from "..";
 import { ChartPanel } from "../../components/chart-panel";
 import { ChartPreview } from "../../components/chart-preview";
-import { DataSetHint } from "../../components/hint";
-import {
-  DataCubeTheme,
-  DataCubeResultOrder,
-  useDataCubesQuery,
-  DataCubeOrganization,
-} from "../../graphql/query-hooks";
-import { useLocale } from "../../src";
 import { ChartConfiguratorTable } from "../table/table-chart-configurator";
 import { ChartAnnotationsSelector } from "./chart-annotations-selector";
 import { ChartAnnotator } from "./chart-annotator";
 import { ChartConfigurator } from "./chart-configurator";
 import { ChartOptionsSelector } from "./chart-options-selector";
 import { ChartTypeSelector } from "./chart-type-selector";
-import { DataSetMetadata } from "./dataset-metadata";
-import { DataSetPreview } from "./dataset-preview";
-import {
-  SearchDatasetBox,
-  SearchFilters,
-  DatasetResults,
-  useSearchQueryState,
-} from "./dataset-search";
 import {
   PanelLeftWrapper,
   PanelMiddleWrapper,
   PanelRightWrapper,
 } from "./layout";
+import { SelectDatasetStep } from "./SelectDatasetStep";
 import { Stepper } from "./stepper";
 
-const TextLink = (props: LinkProps) => {
+export const TextLink = (props: LinkProps) => {
   return (
     <Link
       color="monochrome700"
@@ -46,89 +29,6 @@ const TextLink = (props: LinkProps) => {
       }}
       {...props}
     />
-  );
-};
-
-const SelectDatasetStep = () => {
-  const locale = useLocale();
-
-  const searchQueryState = useSearchQueryState();
-  const { query, order, includeDrafts, filters } = searchQueryState;
-
-  const [state, dispatch] = useConfiguratorState();
-  const [debouncedQuery] = useDebounce(query, 150, {
-    leading: true,
-  });
-
-  // Use the debounced query value here only!
-  const [{ fetching, data }] = useDataCubesQuery({
-    variables: {
-      locale,
-      query: debouncedQuery,
-      order,
-      includeDrafts,
-      filters: filters
-        ? filters.map((filter) => {
-            return { type: filter.__typename, value: filter.iri };
-          })
-        : [],
-    },
-  });
-
-  if (state.state !== "SELECTING_DATASET") {
-    return null;
-  }
-  return (
-    <>
-      <PanelLeftWrapper>
-        <SearchFilters searchQueryState={searchQueryState} />
-      </PanelLeftWrapper>
-      <PanelMiddleWrapper>
-        {state.dataSet ? null : (
-          <Box mb={4}>
-            <SearchDatasetBox
-              searchQueryState={searchQueryState}
-              searchResult={data}
-            />
-          </Box>
-        )}
-        {state.dataSet || !data ? (
-          <>
-            <Box mb={4} px={4}>
-              <TextLink
-                onClick={(ev) => {
-                  ev.preventDefault();
-                  dispatch({
-                    type: "DATASET_SELECTED",
-                    dataSet: undefined,
-                  });
-                }}
-              >
-                Back to the list
-              </TextLink>
-            </Box>
-            <ChartPanel>
-              {state.dataSet ? (
-                <>
-                  <DataSetPreview dataSetIri={state.dataSet} />
-                </>
-              ) : (
-                <DataSetHint />
-              )}
-            </ChartPanel>
-          </>
-        ) : (
-          <ChartPanel>
-            <div>
-              <DatasetResults fetching={fetching} data={data} />
-            </div>
-          </ChartPanel>
-        )}
-      </PanelMiddleWrapper>
-      <PanelRightWrapper>
-        {state.dataSet ? <DataSetMetadata dataSetIri={state.dataSet} /> : null}
-      </PanelRightWrapper>
-    </>
   );
 };
